Move HDF5 loggers to axion.logging; delete dead wrapper modules
The three persistent-state loggers (SimulationHDF5Logger,
DatasetHDF5Logger, AdjointHDF5Logger) lived in src/axion/core/ for
historical reasons but are an orthogonal observability concern, not
solver code. Moved them into the existing axion.logging package, which
previously held an unused generic HDF5 wrapper (HDF5Logger / NullLogger
/ HDF5Reader — no caller anywhere in the tree, confirmed by grep).

LoggingConfig stays in axion.core.logging_config since it is a
sub-config of the engine config tree and is consumed by the engine
constructor like any other sub-config. core/base_engine.py now imports
the loggers from axion.logging; the configs continue to come from
core.logging_config.

The moved files use absolute imports against axion.core to avoid the
core ↔ logging cycle being load-order-sensitive. Smoke test:
obstacle_benchmark with logging=hdf5 writes the expected HDF5 file end
to end.
