+++ src/axion/logging/__init__.py
@@ -1,6 +1,17 @@
-from .hdf5_logger import HDF5Logger
-from .hdf5_reader import HDF5Reader
-from .null_logger import NullLogger
+"""Persistent-state logging for the Axion engine.
+
+Three domain-specific HDF5 loggers, each consumed by AxionEngineBase
+when the corresponding sub-config (``LoggingConfig.{hdf5,dataset,adjoint}``)
+is enabled. The configs themselves live in ``axion.core.logging_config``
+because they are part of the engine config tree.
+"""
+from .adjoint_logger import AdjointHDF5Logger
+from .dataset_logger import DatasetHDF5Logger
+from .sim_logger import SimulationHDF5Logger
 
 
-__all__ = ["HDF5Logger", "HDF5Reader", "NullLogger"]
+__all__ = [
+    "AdjointHDF5Logger",
+    "DatasetHDF5Logger",
+    "SimulationHDF5Logger",
+]
